genereach_aoi/function_test/requirement.txt
numpy==1.23.2
Pillow==9.2.0
--extra-index-url https://download.pytorch.org/whl/lts/1.8/cpu
torch==1.8.2
torchvision==0.9.2